Using LibMultiSense v6.0.0. Added aux camera config

+++ src/Scripts/Objects/MultiSenseCamera.cpp
@@ -140,7 +140,7 @@ void MultiSenseCamera::update() {
     d2->exposure = ptr->exposure;
     d2->scaleIBLAmbient = ptr->scaleIBLAmbient;
     d2->debugViewInputs = ptr->debugViewInputs;
-    d2->prefilteredCubeMipLevels = static_cast<float>(renderUtils.skybox.prefilteredCubeMipLevels);
+    d2->prefilteredCubeMipLevels =renderUtils.skybox.prefilteredCubeMipLevels;
 
 
 }
